Move lazyproperty to conda.cli.actions
